# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WEB API",
        "title": "Rent Transport",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/Rent/:rentId": {
            "get": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Get rent by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "GetRent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "rent id",
                        "name": "rentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/Rent/End/:rentId": {
            "post": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "End renting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "End",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "rent id",
                        "name": "rentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "latitude of tranposrt",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "longitude of tranposrt",
                        "name": "long",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/Rent/MyHistory": {
            "get": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Get the user history of rents",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "MyHistory",
                "responses": {}
            }
        },
        "/api/Rent/New/:transportId": {
            "post": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Start renting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "StartRent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tranport id",
                        "name": "transportId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "type of renting: minutes or days",
                        "name": "rentType",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/Rent/Transport": {
            "get": {
                "description": "Get transport which not far than radius from point",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "GetTransport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "latitude of point",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "longitude of point",
                        "name": "long",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "radius",
                        "name": "radius",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "type of transport",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/Rent/TransportHistory/:trnaposrtId": {
            "get": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Get history of transport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "TranposrtHistory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "transport id",
                        "name": "transportId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/Transport": {
            "post": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Create transport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transport"
                ],
                "summary": "CreateTransport",
                "parameters": [
                    {
                        "description": "Transport",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Transport"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/Transport/:id": {
            "get": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Get information about transport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transport"
                ],
                "summary": "TransportInformation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transport id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Edit the Transport information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transport"
                ],
                "summary": "UpdateTransport",
                "parameters": [
                    {
                        "description": "Transport",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Transport"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Transport id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Remove the Transport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transport"
                ],
                "summary": "DeleteTransport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transport id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "entities.Transport": {
            "type": "object",
            "properties": {
                "can_be_rented": {
                    "type": "boolean"
                },
                "color": {
                    "type": "string"
                },
                "day_price": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "minute_price": {
                    "type": "number"
                },
                "model": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "integer"
                },
                "transport_type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}